+++ lib/utils/utils.dart
@@ -109,6 +109,15 @@ class StringUtils {
     return fullName;
   }
 
+  static String truncate({required String value, required int atLength}) {
+    int valueLength = value.length;
+    if ((atLength > 0) && (valueLength > atLength)) {
+      return '${value.substring(0, value.substring(0, atLength).lastIndexOf(" "))} ...';
+    } else {
+      return value;
+    }
+  }
+
   /// US Phone validation  https://github.com/rokwire/illinois-app/issues/47
 
   static const String _usPhonePattern1 = "^[2-9][0-9]{9}\$";          // Valid:   23456789120